+++ src/loki.js
@@ -2,33 +2,41 @@ const log = require('evillogger')({ ns:'loki' });
 const path = require('path');
 const fs = require('fs-extra');
 const klawSync = require('klaw-sync');
-const loki = require('lokijs');
+const async = require('async');
 
 const config = require('./config');
 const shared = require('./methods/shared');
 
-function initialize() {
+function initialize(callback) {
 
     if (!fs.pathExistsSync(config.SLOKI_DIR_DBS)) {
         fs.ensureDirSync(config.SLOKI_DIR_DBS);
         log.info(`directory ${config.SLOKI_DIR_DBS} created`);
     }
 
-    let dbName;
     let file;
 
+    log.info(`loading databases from ${config.SLOKI_DIR_DBS}`);
+
+    const dbs = [];
+
     for (file of klawSync(config.SLOKI_DIR_DBS)) {
-        dbName = path.basename(file.path).replace(/\.json/, '');
-        log.info(`loading database ${file.path}`);
-        shared.dbs[dbName] = new loki(file.path, shared.DEFAULT_DATABASE_OPTIONS);
+        if (file.path.match(/\.(json|db)$/)) {
+            dbs.push(path.basename(file.path).replace(/\.(json|db)/, ''));
+        }
     }
 
-    const dbTestFile = path.resolve(config.SLOKI_DIR_DBS+'/test.json');
-
-    if (!shared.dbs['test']) {
-        shared.dbs['test'] = new loki(dbTestFile, shared.DEFAULT_DATABASE_OPTIONS);
-        shared.dbs['test'].save();
+    if (dbs.indexOf('test')<0) {
+        dbs.push('test');
     }
+
+    async.each(dbs, (db, next) => {
+        log.info(`loading database ${db}`);
+        shared.createDatabase(db, () => {
+            log.info(`database ${db} loaded`);
+            next();
+        });
+    }, callback);
 }
 
 module.exports = {