+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 ### Changed
 * switch to named parameter in jsonrpc layer [#20](https://github.com/sloki-project/sloki/issues/20)
-* no response needed if attribute "nr" exists in params (lazy mode)
+* switch to switch to loki-fs-structured-adapter by default [#32](https://github.com/sloki-project/sloki/issues/32)
 * listen on multiple port per transport/protocol by default [#25](https://github.com/sloki-project/sloki/issues/25)